Fix and clarify Docker USE_GIT flag

+++ docker/build-database.sh
@@ -14,14 +14,17 @@ set -e
 : ${DBSETTINGS_PATH:=settings.rb}
 : ${DBDATA_CLEANUP:=true} # A reason not to clean it up would be to cache the files using buildkit
 
-# If the source code isn't made available by the extending Dockerfile we need to use git to check it out
-USE_GIT=! [[ -d "${DBSOURCE_PATH}" ]]
-
-[[ ${USE_GIT} ]] && echo 'Source code present in GIT..'
-! [[ ${USE_GIT} ]] && echo 'Source code present locally..'
+# If the source code isn't made available by the extending Dockerfile we need to clone it.
+if [[ -d "${DBSOURCE_PATH}" ]]; then
+  CLONE_DBSOURCE=false
+  echo 'Source code present locally..'
+else
+  CLONE_DBSOURCE=true
+  echo 'Source code present in GIT..'
+fi
 
 # Do git specific validations and fix paths if needed
-if [[ ${USE_GIT} ]]; then
+if [[ "${CLONE_DBSOURCE}" == true ]]; then
   : ${GIT_USERNAME?'GIT_USERNAME must be provided'}
   : ${GIT_TOKEN?'GIT_TOKEN must be provided'}
   : ${GIT_HOSTNAME?'GIT_HOSTNAME must be provided'}
@@ -34,19 +37,20 @@ if [[ ${USE_GIT} ]]; then
   DBCONFIG_PATH="${GIT_REPOSITORY}/${DBCONFIG_PATH}"
 else
   # We require a version outside of git
-   : ${DATABASE_VERSION?'DATABASE_VERSION must be set if source code is present locally..'}
+  : ${DATABASE_VERSION?'DATABASE_VERSION must be set if source code is present locally..'}
 fi
 
-# git is needed to clone the product repo (when USE_GIT) and always for clean external-module clones
+# Install git and ssh
 apk --no-cache add --virtual .git-deps git openssh
+git --version
+ssh -V
 
 # Make our own PGDATA.
 # We are unfortunately doing this because we want the data to persist but the default PGDATA directory is marked as a volume, which cannot be undone.
 mkdir -p "${PGDATA}" && chown -R postgres:postgres "${PGDATA}" && chmod 777 "${PGDATA}"
 
 # fetch repo if needed
-[[ ${USE_GIT} ]] && git --version
-[[ ${USE_GIT} ]] && git clone "https://${GIT_USERNAME}:${GIT_TOKEN}@${GIT_HOSTNAME}/${GIT_ORG}/${GIT_REPOSITORY}.git"
+[[ "${CLONE_DBSOURCE}" == true ]] && git clone "https://${GIT_USERNAME}:${GIT_TOKEN}@${GIT_HOSTNAME}/${GIT_ORG}/${GIT_REPOSITORY}.git"
 
 # create db-data folder for the repo
 mkdir -p "${DBDATA_PATH}"
@@ -110,11 +114,9 @@ else
   pg_resetwal --pgdata "${PGDATA}"
 fi
 
-# image cleanup (removing unneeded db-data, git directory and git dependencies)
-if [[ "${DBDATA_CLEANUP}" == 'true' ]]; then
-  rm -rf "${DBDATA_PATH}"
-fi
-[[ ${USE_GIT} ]] && rm -rf "${GIT_REPOSITORY}"
+# Image cleanup (removing unneeded db-data, git directory and git dependencies)
+[[ "${DBDATA_CLEANUP}" == 'true' ]] && rm -rf "${DBDATA_PATH}"
+[[ "${CLONE_DBSOURCE}" == true ]] && rm -rf "${GIT_REPOSITORY}"
 apk del .git-deps
 # Exit with 0 if this stage is reached, otherwise the return code from
 #  the last if statement might be used, which might let Docker think the build failed
